=Seperated Module for CURL and Made webhookmessage function async fixing lag issue when used in takeover event
