+++ src/build.sh
@@ -3,7 +3,7 @@
 # E.g.: ./build.sh -d --sqlite
 
 cc="g++"
-options="-I. -m32 -fPIC -Wall -fvisibility=hidden"
+options="-I. -m32 -fPIC -Wall -fvisibility=hidden -std=c++11 -pthread"
 
 separator="---------------------"
 
@@ -130,6 +130,11 @@ $cc $debug $options $constants -c libcod.cpp -o objects/libcod.opp
 
 echo "##### COMPILE QVSNPRINTF.C #####"
 $cc $debug $options $constants -c vendor/qvsnprintf.c -o objects/qvsnprintf.opp
+
+if [ $libcurl_found == 1 ]; then
+    echo "##### COMPILE GSC_CURL.CPP #####"
+    $cc $debug $options $constants -c gsc_curl.cpp -o objects/gsc_curl.opp
+fi
 
 echo "##### LINK    libcod1.so #####"
 objects="$(ls objects/*.opp)"
