+++ src/gsc_curl.hpp
@@ -1,0 +1,9 @@
+#ifndef _GSC_CURL_HPP_
+#define _GSC_CURL_HPP_
+
+#include "gsc.hpp"
+
+#if COMPILE_LIBCURL == 1
+void gsc_curl_webhookmessage();
+#endif
+#endif
