+++ src/gsc_utils.hpp
@@ -23,8 +23,4 @@ void gsc_utils_makelocalizedstring();
 void gsc_utils_ban();
 void gsc_utils_unban();
 
-#if COMPILE_LIBCURL == 1
-void gsc_utils_webhookmessage();
-#endif
-
 #endif
